% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ComBatv3.34}
\alias{ComBatv3.34}
\title{ComBat v3.34 batch correction}
\usage{
ComBatv3.34(
  dat,
  batch,
  mod = NULL,
  par.prior = TRUE,
  prior.plots = FALSE,
  mean.only = FALSE,
  ref.batch = NULL,
  BPPARAM = bpparam("SerialParam")
)
}
\arguments{
\item{dat}{}

\item{batch}{}

\item{mod}{}

\item{par.prior}{}

\item{prior.plots}{}

\item{mean.only}{}

\item{ref.batch}{}

\item{BPPARAM}{}
}
\value{

}
\description{
Previous version of ComBat (sva version 3.34.0) This is for
compatibility with the replication of SCoPE2 (version2) and
zhu2019EL.
}
