% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correctIsotopicCarryover}
\alias{correctIsotopicCarryover}
\title{Isotopic carryover correction for plexDIA data}
\usage{
correctIsotopicCarryover(x)
}
\arguments{
\item{x}{An object that inherits from the \code{SummarizedExperiment}
class.}
}
\value{
An object of same class as \code{x} with the isotopic correction
applied.
}
\description{
This function performs isotopic carryover correction. Currently,
only isotopic correction for mTRAQ is available.
}
