% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{countUniqueFeatures}
\alias{countUniqueFeatures}
\title{Count Unique Features}
\usage{
countUniqueFeatures(object, i, rowDataCol = NULL, varName = "count")
}
\arguments{
\item{object}{}

\item{i}{}

\item{rowDataCol}{}

\item{varName}{}
}
\value{

}
\description{
Description of the function
}
