% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggalusGeneDf}
\alias{ggalusGeneDf}
\title{Galus Galus protein ID to protein symbol table}
\format{
An object of class \code{data.frame} with 660 rows and 2 columns.
}
\usage{
ggalusGeneDf
}
\description{
A \code{data.frame} with 2 columns: one that contains Ensembl protein IDs and
the other containing the protein symbols for the proteins selected by the
zhu2019EL dataset
}
\keyword{datasets}
