% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longRowData}
\alias{longRowData}
\title{longRowData}
\usage{
longRowData(object, i, rowDataCols)
}
\arguments{
\item{object}{An object of class QFeatures.}

\item{i}{Index of the assay of interest.}

\item{rowDataCols}{The row data column names.}
}
\value{
A RowData DataFrame in long format.
}
\description{
longRowData
}
\examples{

print("TODO")
}
