% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replaceRowDataCols}
\alias{replaceRowDataCols}
\title{Replace columns in the rowData}
\usage{
replaceRowDataCols(object, i, col, value)
}
\arguments{
\item{object}{An object of class \code{QFeatures}}

\item{i}{A \code{numeric()} or \code{character()} vector indicating from
which assays the \code{rowData} should be taken.}

\item{col}{A \code{numeric()} or \code{character()} vector indicating the
column(s) to replace.}

\item{value}{A \code{list()} of same length as \code{i}. If \code{i} is provided
as a character vector, the elements of \code{value} must be named
after \code{i}. Each element should contain the replacement values to insert in the rowData. If
\code{length(col) > 1}, each element should be a matrix with the
same rows as its corresponding assay and the same number of
columns as the length of col.}
}
\value{
An object of class \code{QFeatures}
}
\description{
The function replaces columns in the rowData of assays from a
\code{QFeatures} object. This function can also be used to remove column
in the rowData (see examples).
}
\examples{
data("feat2")
rowData(feat2)[[1]]
## Add a new column to all assay
values <- lapply(rowData(feat2), function(x) rep("bar", nrow(x)))
feat2 <- replaceRowDataCols(feat2, 1:3, "foo", values)
rowData(feat2)[[1]]
## Update column
values <- lapply(rowData(feat2), function(x) rep("barNew", nrow(x)))
feat2 <- replaceRowDataCols(feat2, 1:3, "foo", values)
rowData(feat2)[[1]]
## Remove column
values <- lapply(1:3, function(x) NULL)
feat2 <- replaceRowDataCols(feat2, 1:3, "foo", values)
rowData(feat2)[[1]]
}
