% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scp_imputeKNN}
\alias{scp_imputeKNN}
\title{KNN imputation (for replication)}
\usage{
scp_imputeKNN(obj, i, name = "KNNimputedAssay", k = 3)
}
\arguments{
\item{obj}{}

\item{i}{}

\item{name}{}

\item{k}{}
}
\value{

}
\description{
KNN imputation
}
