% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longRowData}
\alias{longRowData}
\title{longRowData}
\usage{
longRowData(object, i, rowDataCols)
}
\arguments{
\item{object}{}

\item{i}{}

\item{rowDataCols}{}
}
\value{

}
\description{
longRowData
}
\examples{

print("TODO")
}
